requirements.txt
absl-py==2.3.1
agents==1.4.0
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
certifi==2025.10.5
charset-normalizer==3.4.4
cloudpickle==3.1.1
colorama==0.4.6
DateTime==5.5
flatbuffers==25.9.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0
gym==0.26.2
gym-notices==0.1.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.11.3
langchain==1.0.2
langchain-core==1.0.1
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.38
libclang==18.1.1
loguru==0.7.3
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
numpy==2.3.4
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
pillow==12.0.0
protobuf==6.33.0
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
win32_setctime==1.2.0
wrapt==2.0.0
xxhash==3.6.0
zope.interface==8.0.1
zstandard==0.25.0